Split off tree into its own thing.
